Fix nodebatcher to be both simpler and more efficient
